requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-aiplatform==1.38.1
google-cloud-pubsub==2.18.4
google-cloud-run==0.10.3
google-generativeai==0.3.2
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0